Refactor: Combatant requires a common interface for select_attack() and attack()
