+++ common.py
@@ -5,6 +5,7 @@ This module should not import other game modules.
 """
 
 import json
+from typing import Sequence
 
 
 #Status
@@ -15,6 +16,38 @@ with open("statuses.json", "r") as f:
         for record in json.load(f)
     }
     statuses = list(statusdata.keys())
+
+
+class Attack:
+    """Encapsulates data for an attack."""
+    def __init__(self,
+                name: str,
+                accuracy: "int | None" = None,
+                damage: "int | None" = None,
+                healing: "int | None" = None,
+                repeats: Sequence[int] = (1, 1),
+                inflicts: "str | None" = None):
+        assert len(repeats) == 2
+        min_, max_ = repeats
+        assert isinstance(min_, int)
+        assert isinstance(max_, int)
+        assert min_ >= 0
+        assert max_ >= min_
+        self.name = name
+        self.accuracy = accuracy
+        self.damage = damage
+        self.healing = healing
+        self.inflicts = inflicts
+        self.repeats = tuple(repeats)
+
+    def __repr__(self) -> str:
+        return f"Attack({self.name!r}, accuracy={self.accuracy}, damage={self.damage}, repeats={self.repeats!r}, inflicts={self.inflicts!r})"
+
+    def __str__(self) -> str:
+        if self.repeats:
+            lower, upper = self.repeats
+            return f"{self.name}  {self.accuracy} acc  {self.damage}*n dmg, n ranges from {lower} to {upper}"
+        return f"{self.name}  {self.accuracy} acc  {self.damage} dmg"
 
 
 class Status:
@@ -90,6 +123,16 @@ class Combatant:
     def has_status(self, status):
         """If character has status, returns True, else returns False."""
         return self.status[status] is not None
+
+    def select_attack(self) -> Attack:
+        """Select an attack to use in combat"""
+        raise NotImplementedError
+
+    def attack(self, target: "Combatant", attack: Attack) -> bool:
+        """Attacks a target using one of its attacks.
+        Returns True if the attack succeeded, otherwise False.
+        """
+        raise NotImplementedError
 
     def update(self) -> None:
         """Updates character state at end of turn."""
